pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcpisia"
version = "0.1.0"
description = "MCP implementation with GitHub Agent Orchestrator"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "MCPISIA Team"},
]
keywords = ["mcp", "github", "ai", "claude", "orchestration", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v2 (GPLv2)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # MCP Core
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "aiofiles>=23.0.0",
    "redis>=4.5.0",
    "typing-extensions>=4.5.0",
    "python-dotenv>=1.0.0",
    "structlog>=23.0.0",
    # GitHub Agent Orchestrator  
    "anthropic>=0.21.0",
    "PyGithub>=1.59.0",
    "aiohttp>=3.8.0",
    "click>=8.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "black>=23.0.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/eagurin/MCP-"
Repository = "https://github.com/eagurin/MCP-.git"
Issues = "https://github.com/eagurin/MCP-/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
mcpisia-server = "src.server:main"
github-orchestrator = "src.main:main"
agent-cli = "agent_cli:main"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
